refactor: loading 컴포넌트로 추상화
