--- src/components/loadingContainer/index.tsx
@@ -1,8 +1,0 @@
-import { ReactNode } from "react";
-import { Container } from "./index.style";
-
-const LoadingContainer = ({ children }: { children: ReactNode }) => {
-  return <Container>{children}</Container>;
-};
-
-export default LoadingContainer;
