+++ src/pages/diagnosis/index.style.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { Heading_3 } from "src/lib/fontStyle";
+import { Heading_3, Body_1, Body_3 } from "src/lib/fontStyle";
 
 export const Container = styled.section`
   height: calc(100vh - 5.6rem);
@@ -27,4 +27,38 @@ export const Question = styled(Heading_3)`
 
   margin-top: 7rem;
   padding: 0 5rem;
+`;
+
+export const LoadingTitle = styled(Heading_3)`
+  font-weight: 200;
+  text-align: center;
+  word-break: keep-all;
+  color: ${({ theme }) => theme.color.grey_200};
+
+  .highlight {
+    font-weight: 500;
+  }
+`;
+
+export const LoadingIcon = styled.img`
+  width: 26rem;
+  height: 24.8rem;
+`;
+
+export const LoadingBottomText = styled.section`
+  text-align: center;
+`;
+
+export const Tips = styled(Body_3)`
+  font-weight: 200;
+  letter-spacing: -0.05rem;
+
+  color: ${({ theme }) => theme.color.grey_300};
+`;
+
+export const Description = styled(Body_1)`
+  margin-top: 0.8rem;
+  width: 17.2rem;
+
+  color: ${({ theme }) => theme.color.grey_400};
 `;
